refactor(edge): reduce bootstrap tls module to compatibility facade
