+++ crates/edge/src/quic_listener/bootstrap/mod.rs
@@ -1,6 +1,7 @@
 mod context;
 mod dispatch;
 mod intake;
+mod listener;
 mod outcome;
 mod request;
 mod response;
@@ -8,31 +9,8 @@ mod startup;
 mod state;
 mod websocket;
 
-pub(in crate::quic_listener) use self::context::{
-    BootstrapDispatchCtx, BootstrapRequestCtx, BootstrapRuntimeCtx,
-};
-pub(in crate::quic_listener) use self::dispatch::{
-    BootstrapDispatchInput, dispatch_bootstrap_upstream,
-};
-pub(in crate::quic_listener) use self::intake::{
-    BootstrapRequestIntake, bootstrap_error_response, prepare_bootstrap_request_intake,
-};
-pub(in crate::quic_listener) use self::outcome::observe_bootstrap_request_proxy_error;
-pub(in crate::quic_listener) use self::request::{
-    BootstrapBuildRequestInput, BootstrapPolicyEvaluationInput, BootstrapPreparedRoute,
-    build_bootstrap_upstream_request, evaluate_bootstrap_request_policy,
-};
-pub(in crate::quic_listener) use self::response::{
-    BootstrapStreamingBody, BootstrapWritebackInput, boxed_full, write_bootstrap_response,
-};
-pub(in crate::quic_listener) use self::startup::{
-    PreparedBootstrapListenerStartup, prepare_bootstrap_listener_startup,
-    spawn_bootstrap_listener_task,
-};
+pub(in crate::quic_listener) use self::listener::spawn_bootstrap_tls_listener;
+#[cfg(test)]
 pub(in crate::quic_listener) use self::state::{
     BootstrapConnectionState, BootstrapStartupState, bootstrap_connection_state,
-};
-pub(in crate::quic_listener) use self::websocket::{
-    capture_bootstrap_websocket_flow, dispatch_bootstrap_websocket,
-    write_bootstrap_websocket_upgrade,
 };
